CSP-1552 - refactored to address PR comments
