+++ src/SFA.DAS.Provider.PR.Web/AppStart/ConfigurationKeys.cs
@@ -1,0 +1,9 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace SFA.DAS.Provider.PR.Web.AppStart;
+
+[ExcludeFromCodeCoverage]
+public static class ConfigurationKeys
+{
+    public const string EncodingConfig = "SFA.DAS.Encoding";
+}
