+++ src/SFA.DAS.Provider.PR.Web/Program.cs
@@ -27,8 +27,6 @@ builder.Services
         builder.AddFilter<ApplicationInsightsLoggerProvider>("Microsoft", LogLevel.Information);
     })
     .AddHttpContextAccessor()
-    .AddConfiguration(configuration)
-    .AddEncodingService()
     .AddServiceRegistrations(configuration)
     .AddSession(configuration)
     .AddHealthChecks(configuration)
